Begin MVVM refactorization
