+++ app/build.gradle
@@ -11,11 +11,16 @@ android {
         versionCode 6
         versionName "1.05"
     }
+
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
+    }
+
+    dataBinding {
+        enabled = true;
     }
 }
