+++ app/src/main/AndroidManifest.xml
@@ -13,11 +13,11 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".Activities.MainActivity"
+            android:name=".View.MainActivity"
             android:configChanges="orientation|keyboardHidden"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".Activities.InitActivity"
+            android:name=".View.InitActivity"
             android:configChanges="orientation|keyboardHidden"
             android:screenOrientation="portrait"
             android:theme="@style/SplashTheme">
@@ -28,7 +28,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".Activities.DormActivity"
+            android:name=".View.DormActivity"
             android:configChanges="orientation|keyboardHidden"
             android:screenOrientation="portrait" />
         <receiver
