+++ app/src/main/java/io/ericlee/illinilaundry/View/MainActivity.java
@@ -1,4 +1,4 @@
-package io.ericlee.illinilaundry.Activities;
+package io.ericlee.illinilaundry.View;
 
 import android.graphics.Color;
 import android.support.design.widget.TabLayout;
