+++ app/src/main/res/layout/activity_init.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".Activities.InitActivity"
+    tools:context=".View.InitActivity"
     android:background="#002058">
 
     <ImageView
